Move bounded type definitions to sp-runtime (#11645)
* Move bounded type definitions to sp-runtime

* cargo fmt

* Fix compile error


* Move TryCollect to sp-runtime

* Write some docs

* Import missing types
